requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.5
django-admin-interface==0.14.2
django-braces==1.14.0
django-colorfield==0.3.2
django-debug-toolbar==3.2
django-embed-video==1.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-memcache-status==2.2
django-suit @ https://github.com/darklow/django-suit/tarball/develop
djangorestframework==3.12.2
easy-thumbnails==2.7.1
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
Pillow==8.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-memcached==1.59
pytz==2020.5
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted @ file:///E:/Vs%20Code%20Python/django%20project/bookfourth/Twisted-20.3.0-cp39-cp39-win_amd64.whl
txaio==20.12.1
urllib3==1.26.2
zope.interface==5.2.0
